renderer/renderer_impl: Add missing float to int casting

+++ src/renderer/renderer_impl.cpp
@@ -256,8 +256,8 @@ void RendererImpl::AdjustCaptionArea(int origin_plane_width, int origin_plane_he
     float y_magnification = static_cast<float>(video_area_height_) / static_cast<float>(origin_plane_height);
     float magnification = std::min(x_magnification, y_magnification);
 
-    int caption_area_width = std::floor(static_cast<float>(origin_plane_width) * magnification);
-    int caption_area_height = std::floor(static_cast<float>(origin_plane_height) * magnification);
+    int caption_area_width = static_cast<int>(std::floor(static_cast<float>(origin_plane_width) * magnification));
+    int caption_area_height = static_cast<int>(std::floor(static_cast<float>(origin_plane_height) * magnification));
     int caption_area_start_x = (video_area_width_ - caption_area_width) / 2;
     int caption_area_start_y = (video_area_height_ - caption_area_height) / 2;
